api/requirements.txt
google-cloud-logging==3.14.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
google-cloud-bigquery==3.27.0
google-cloud-firestore==2.19.0
firebase-admin==6.5.0
